WIP: srgb egl and x11
This commit makes EGL respect the source buffers being srgb,
which makes the alpha blending example from the referenced issue
appear correct, but the text weight is impacted somehow.

I'm also not happy with the outputs_srgb parameter on the shaders:
I have to turn it off to prevent over-correcting, and I'm pretty
sure that won't work out on other platforms.

More experimentation is needed, so this is going in a different
branch until I get things sorted out.

refs: https://github.com/wez/wezterm/issues/544