gui: add glScissor to cut out 1-physical-px border on Wayland HiDPI
With increment step resize, we're rounding up to cell size in *physical* pixels,
but the surface has to have an integer logical size
and the buffer has to be *that* times scale factor.

So when the rounded size has half-pixels, there would be an extra 1px (on 2x scale)
gap displayed on the screen. To avoid rendering it, use scissor.