Add full assignment example. Clarify Shift v SHIFT.
The existing example uses "Nop" which can be supplied as string but other key assignments require use of `wezterm.action`.

For people not familiar with Lua (like me :) ) a full example with correct syntax is helpful at this point in the documentation.

Additionally, it was not immediately obvious to me that the Shift key requires different capitalization in different contexts. The resulting error messages also don't mention the significance of the capitalization. So this commit adds a note about this detail.

Hopefully these additions will make the initial key assignment experience more approachable & less frustrating. :)

(The key combination used in this example matches that used by the elementary OS terminal.)