mux: fix local/remote tab tracking
The issue here is that when spawning a new tab in the mux client,
we didn't record the local/remote tab id mapping.

Subsequently, when processing the overall list of tabs, we'd realize
that the mapping was missing and allocate a new local tab to act
as the holder for the existing remote tab and end up with a strange
aliased view of the tab in a second tab!

This commit establishes that mapping at the time that we spawn
the new pane.

refs: https://github.com/wez/wezterm/issues/1222