ci: switch to GH CLI for uploading to releases
I lost a few hours over the weekend because the GH release uploads
are flakey and the action I was using doesn't internally perform
retries.  I had to manually delete the failed uploads from the release
and then re-trigger the builds across several platforms, several times
for both of the releases I pushed this weekend :-/

This commit speculatively switches over to using the GH cli in the
hopes that the error reporting is better, and also because it is
simpler to externally drive a retry loop.

Let's see how this goes.